Fix character replacements to use ChrW, don't replace ellipses in CJK
The previous commits from #2 fix files being interpreted/converted in
the wrong encoding in some cases. This means the files are now always
interpreted as UTF-8 (as they should be) and then converted to UTF-16
for VBA's strings. This means character replacements should no longer
be done as raw UTF-8 byte sequences, but as UTF-16 ones, which ChrW()
can help us to get. This makes the code more readable as well, yay :)
Turns out our UTF-8 byte sequences were actually 16-bit for a byte...

Also, the sanitization step to convert `…` to `...` only makes sense
for languages which use the 8x8 font - the sanitization is done because
Excel unwantedly converts three dots to a single ellipsis character,
but CJK languages actually intentionally use it. So now we only replace
ellipsis characters if the font is something other than 8x8.